45: Added ConnectedProfileWidget subcomponent
